+++ Src/WindowsTerminal/WindowsTerminal.cs
@@ -3,8 +3,8 @@
 // This code is licensed under MIT license (see LICENSE for details)
 // --------------------------------------------------------------------------
 
+using System.Text;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 
 using Webmaster442.WindowsTerminal.Internals;
 
@@ -15,22 +15,50 @@ namespace Webmaster442.WindowsTerminal;
 /// </summary>
 public static class WindowsTerminal
 {
-    private static readonly string _localFragments = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "Microsoft", "Windows Terminal", "Fragments");
+    private static readonly string LocalFragments = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "Microsoft", "Windows Terminal", "Fragments");
 
-    private static readonly JsonSerializerOptions _serializerOptions = CreateOptions();
-
-    private static JsonSerializerOptions CreateOptions()
+    private static string GetControlSequenceResponse(string controlSequence)
     {
-        var options = new JsonSerializerOptions
+        var response = new StringBuilder();
+
+        try
         {
-            WriteIndented = true,
-            PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingDefault,
-            TypeInfoResolver = TerminalFragmentGenerationContext.Default
-        };
-        options.Converters.Add(new JsonStringEnumConverter<TerminalBackgroundImageAlignment>(JsonNamingPolicy.CamelCase));
-        options.Converters.Add(new JsonStringEnumConverter<TerminalBackgroundImageStretchMode>(JsonNamingPolicy.CamelCase));
-        return options;
+            Console.Write(controlSequence);
+            Thread.Sleep(20);
+            while (Console.KeyAvailable)
+            {
+                char? c = Console.ReadKey(true).KeyChar;
+                response.Append(c);
+            }
+            return response.ToString();
+        }
+        catch (IOException)
+        {
+            return string.Empty;
+        }
+    }
+
+    /// <summary>
+    /// Get the specified color of the current terminal palette
+    /// </summary>
+    /// <param name="colorIndex">Color index to get. Must be between 0 and 15</param>
+    /// <returns>24 bit RGB color</returns>
+    public static (byte r, byte g, byte b) GetPaletteColor(int colorIndex)
+    {
+        ArgumentOutOfRangeException.ThrowIfLessThan(colorIndex, 0);
+        ArgumentOutOfRangeException.ThrowIfGreaterThan(colorIndex, 15);
+        string response = GetControlSequenceResponse($"\u001B]4;{colorIndex};?\u0007");
+        if (string.IsNullOrEmpty(response))
+            return (0, 0, 0);
+
+        var parts = response.Split([':', ';', '/', '\u001b'], StringSplitOptions.RemoveEmptyEntries);
+
+        if (parts.Length < 6)
+            return (0, 0, 0);
+
+        return (byte.Parse(parts[3][..2], System.Globalization.NumberStyles.HexNumber),
+                byte.Parse(parts[4][..2], System.Globalization.NumberStyles.HexNumber),
+                byte.Parse(parts[5][..2], System.Globalization.NumberStyles.HexNumber));
     }
 
     /// <summary>
@@ -44,7 +72,7 @@ public static class WindowsTerminal
         /// </summary>
         /// <returns>an array of terminal json fragments present in the LocalApplicationData</returns>
         public static string[] GetLocalFragments()
-            => Directory.GetFiles(_localFragments, "*.json", SearchOption.AllDirectories);
+            => Directory.GetFiles(LocalFragments, "*.json", SearchOption.AllDirectories);
 
         /// <summary>
         /// Returns true, if a JSON Fragment extextension is installed
@@ -53,7 +81,7 @@ public static class WindowsTerminal
         /// <param name="fragmentName">fragment json name</param>
         /// <returns>True, if extension is installed</returns>
         public static bool IsFragmentInstalled(string appName, string fragmentName)
-            => File.Exists(Path.Combine(_localFragments, appName, Path.ChangeExtension(fragmentName, ".json")));
+            => File.Exists(Path.Combine(LocalFragments, appName, Path.ChangeExtension(fragmentName, ".json")));
 
         /// <summary>
         /// Try to remove a terminal fragment from the LocalApplicationData
@@ -65,7 +93,7 @@ public static class WindowsTerminal
         {
             try
             {
-                var filePath = Path.Combine(_localFragments, appName, Path.ChangeExtension(fragmentName, ".json"));
+                var filePath = Path.Combine(LocalFragments, appName, Path.ChangeExtension(fragmentName, ".json"));
                 if (File.Exists(filePath))
                 {
                     File.Delete(filePath);
@@ -90,13 +118,13 @@ public static class WindowsTerminal
         {
             try
             {
-                var fragmentFolder = Path.Combine(_localFragments, appName);
+                var fragmentFolder = Path.Combine(LocalFragments, appName);
                 if (!Directory.Exists(fragmentFolder))
                 {
                     Directory.CreateDirectory(fragmentFolder);
                 }
                 var filePath = Path.Combine(fragmentFolder, Path.ChangeExtension(fragmentName, ".json"));
-                using var stream = File.Create(filePath);
+                await using var stream = File.Create(filePath);
                 await JsonSerializer.SerializeAsync(stream, terminalFragment, typeof(TerminalFragment), TerminalFragmentGenerationContext.Default);
                 return true;
             }
@@ -114,16 +142,15 @@ public static class WindowsTerminal
         /// <returns>Terminal fragment data</returns>
         public static async Task<TerminalFragment?> ReadFragmentAsync(string appName, string fragmentName)
         {
-            var filePath = Path.Combine(_localFragments, appName, Path.ChangeExtension(fragmentName, ".json"));
+            var filePath = Path.Combine(LocalFragments, appName, Path.ChangeExtension(fragmentName, ".json"));
             if (!File.Exists(filePath))
             {
                 return null;
             }
-            using var stream = File.OpenRead(filePath);
+            await using var stream = File.OpenRead(filePath);
             return await JsonSerializer.DeserializeAsync<TerminalFragment>(stream, TerminalFragmentGenerationContext.Default.TerminalFragment);
         }
     }
-
 
     /// <summary>
     /// Set the progressbar state
@@ -143,6 +170,13 @@ public static class WindowsTerminal
     /// <returns>True, if App is running inside windows terminal</returns>
     public static bool IsRunningInsideWindowsTerminal()
         => Environment.GetEnvironmentVariable("WT_SESSION") != null;
+
+    /// <summary>
+    /// Get the current session id
+    /// </summary>
+    /// <returns>The current windows Session id. Returns null, if program is not running in Windows Terminal</returns>
+    public static string? GetCurrentSessionId()
+        => Environment.GetEnvironmentVariable("WT_SESSION");
 
     /// <summary>
     /// Set the window title
@@ -184,7 +218,7 @@ public static class WindowsTerminal
         /// The start of the command output / the end of the commandline
         /// </summary>
         public static void CommandExecuted()
-            => Console.Write($"\e]133;C\a");
+            => Console.Write("\e]133;C\a");
 
         /// <summary>
         /// Signal the end of the command